- Re-formatted code.

+++ 001 - SOURCE_CODE/002 - OLIVES_SYSTEM/Olives/ViewModels/EditPatientStatusViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using Shared.Attributes;
-using Shared.Constants;
 using Shared.Enumerations;
 using Shared.Resources;
 
@@ -7,7 +6,7 @@ namespace Olives.ViewModels
 {
     public class EditPatientStatusViewModel : FindPatientViewModel
     {
-        [InNumericArray(new[] {(int)AccountStatus.Inactive, (int)AccountStatus.Active},
+        [InNumericArray(new[] {AccountStatus.Inactive, AccountStatus.Active},
             ErrorMessageResourceType = typeof (Language),
             ErrorMessageResourceName = "InvalidAccountStatus")]
         public byte Status { get; set; }
