+++ 001 - SOURCE_CODE/002 - OLIVES_SYSTEM/Olives/ViewModels/OliveInitializePersonViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using Shared.Attributes;
-using Shared.Constants;
 using Shared.Enumerations;
 using Shared.Resources;
 using Shared.ViewModels;
@@ -11,7 +10,7 @@ namespace Olives.ViewModels
         /// <summary>
         ///     Role of account.
         /// </summary>
-        [InNumericArray(new[] {(int)AccountRole.Patient, (int)AccountRole.Doctor}, ErrorMessageResourceType = typeof (Language),
+        [InNumericArray(new[] {AccountRole.Patient, AccountRole.Doctor}, ErrorMessageResourceType = typeof (Language),
             ErrorMessageResourceName = "InvalidRole")]
         public int Role { get; set; }
     }
